[query] Pin VEP base image os version
We'd like vep results to remain consistent. I've noticed in #14927 when
upgrading hail-ubuntu to 24.04, some results changed. So, we pin our VEP
containers to be based on ubuntu:22.04.

Also here is some minor reorganization to how we group dependencies,
a harmonization of samtools installation across both the 85 and 95
images, and an upgrade to the latest version of samtools.