[query] Pin VEP base image os version (#14929)
We'd like vep results to remain consistent. I've noticed in #14927 when
upgrading hail-ubuntu to 24.04, some results changed, and I'm not sure
why. I'd like to upgrade our images to the newer version. So, I've
pinned our VEP containers to be based on ubuntu:22.04.

Also here is some minor reorganization to how we group dependencies, a
harmonization of samtools installation across both the 85 and 95 images,
and an upgrade to the latest version of samtools.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP