[batch] Log the error if JVM initialization fails due to an exception (#14916)
## Change Description

Fixes #14910.

Followup to PR #14909, this ensures the exception gets logged if the JVM
initialisation task exits early. The previous PR has fixed the main
cause of this task exiting before it has created all the JVMs, but if
there are other failure modes this logging is necessary to both notice
the problem and start to diagnose the cause.

## Security Assessment

- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

- This change has no security impact

### Impact Description

Potentially logs one extra message.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec