[hailtop] Fix tmpdir bug (#14937)
## Change Description

Fixes #14936 

Following
[this](https://github.com/hail-is/hail/pull/14684/files#diff-00fcab89dad87f243267f011f6f2d0edf55b3abb9f0ec77cdcb8d9271627a658L147-L149)
change to remove the trailing '/' from the tmpdir generator, we now need
to add it back in when using the generated tmpdirs.

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

Delete all except the correct answer:
- This change has a low security impact

### Impact Description

Bug fix in the client library to address incorrect formatting of tmpdir
paths.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec