[batch] fix to slow scheduler query (#14948)
## Change Description

The query to get the head of the job queue for a given batch was
building a temp table containing all jobs in the batch. In a batch with
millions of jobs, this was taking >30s, and the scheduler couldn't
schedule jobs fast enough to keep worker instances alive.

Credit to Bram Gorissen for the optimized query. It does change the
order we schedule jobs, and we may want to revisit that ordering, but we
urgently need to remove this bottleneck to complete a running large
batch.

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

Delete all except the correct answer:
- This change has a low security impact

### Impact Description

Only affects the order the batch scheduler considers ready jobs.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec