[qob] Don't set hail off heap memory limits in QoB
Also, set the java heap limit to 90% of machine memory.

The intention here was to give nicer error messages for when hail was
using too much memory. Unfortunately, this drastically reduces the jobs
that will succeed, as we partition the heap into a JVM portion and
a native portion. And so, the job may hit a memory limit while the
machine still has ample memory.