[qob] write any throwable to worker out file (#14977)
This change logs all throwables on the worker during partition execution
to the output.

QoB jobs are failing with FileNotFound exceptions when trying to read
the failed partition logs when the worker dies with OOM. This is because
OOM is considered a fatal error and so we were never writing it to the
output.

This change has no security impact.