[query] Lift backend state into `(Batch|Py4J)QueryDriver` (#14698)
This change extracts common "driver" logic and state out of the various
backend classes into two new classes:
- `BatchQueryDriver`
- `Py4JQueryDriver`

These drivers handle requests from the python clients via their
respective rpc apis. The `BackendRpc` trait implements the core logic of
reading a command, dispatching to the backend for execution, and
responding to the user. Each driver implements the specifics of
- how requests are handled
  - over http in the case of py4j
  - via object storage for batch 
- how execute contexts are created
- how state is managed between requests

With this, driver state and lifecycle becomes decoupled from how queries
are executed (ie which "backend" implementation). It would be trivial to
use a `ServiceBackend` with the `Py4JQueryDriver` to test generated code
on batch without building and uploading a query jar.