[compiler] refactor row/col count estimates (#14921)
## Change Description

This removes another of the arbitrary differences between relational and
value IR classes, preparing to codegen relational IR classes.

The difference being removed here is the implementation of row/col count
upper bounds and exact but optional partition row counts directly in the
IR classes. These are really analysis passes, and should be defined
separately from the node definitions.

To do this cleanly, I use a pattern which I think we should look to use
more often. The goal is to separate the node definitions from the
analysis implementation, so that adding a new node usually doesn't
require modifying or knowing about the analysis.

To achieve this, I add a few traits which an IR class can inherit from,
which consisely express *semantic* properties of the operation. The
analysis then doesn't need to match on specific nodes, instead relying
on this set of properties. The semantic traits are added in
`BaseIR.scala`, and express whether a node preserves rows/cols (each
input row/col corresponds to an output row/col with the same key),
preserves or removes rows/cols (each input row/col is either deleted or
corresponds to an output row/col with the same key), and in both cases
whether it also preserves partitioning (ordering and partition
assignment of rows/cols are preserved), in all cases relative to a
specified child.

The new analysis passes are defined in `analyses/RowCounts.scala`, and
with a few exceptions only match against the new property traits, not
against specific IR classes.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP