[query] don't lower inside CompileAndEvaluate (#14969)
## Change Description

Before this change, `CompileAndEvaluate` always invoked the
`relationalLowerer` lowering pipeline, despite the fact that this is
often called *after* lowering, sometimes on many fragments of a larger
IR. So not only is the lowering pipeline not needed, I believe this is
at least one of the causes of our logs being polluted with so many IRs
in so many stages of lowering it's hard to keep track of what's
happening.

As a step towards untangling the compilation process, this instead makes
`CompileAndEvaluate` take a lowering pipeline as an argument, with a
default of a trivial no-op pipeline. This also allowed the cleanup of
some code duplication.

## Security Assessment

This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP