Add kchandra to AUTHORIZED_USERS (#14991)
## Change Description

Part of https://github.com/hail-is/hail/issues/14971

Adding a new Hail developer to the list of authorized users.

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

Delete all except the correct answer:
- This change has a high security impact

### Impact Description

Replace this content with a description of the impact of the change:
This change gives a new developer admin access, now that security
onboarding is complete.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec