[compiler] fix testBinarySearchOnSet (#14994)
## Change Description

This test was flaky, as scalacheck was discarding the test case whenever
the generated value was not an element of the generated set. Since the
empty set is likely to be generated, this failed often, causing the test
to fail with a "too many discarded cases" error.

There was some logic in the assertion that was clearly intended to test
the behavior in the case that the element is not contained in the set.
But as that logic was never exercised, it was no longer correct.

I've rewritten the assertion to never discard a test case, and to
correctly verify the post-conditions of the binary search method in all
cases. As a side benefit, on my computer, the test has gone from
sometimes taking 30-60s, to <2s.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP