[docker] Use uv managed python in docker images (#14962)
In our 'root' hail-ubuntu container, we now install [uv], a python
package and project manager, and then use it to manage python
installation for our docker images. This change is a step in converting
our build system and python dependency management to entirely use uv.

In the docker image, after installing python, we create a virtual
environment and configure the image to use it by default for all python
operations.

[uv]: https://docs.astral.sh/uv/

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating
- This change has a low security impact

### Impact Description
Changes how we install python away from ubuntu standard package
repositories or (usually) the [deadsnakes] ppa, to using python
distributed by astral, built and distributed in the
[python-build-standalone] project. More details about how these
distributions are built can be found in the [documentation].

[python-build-standalone]:
https://github.com/astral-sh/python-build-standalone
[deadsnakes]: https://launchpad.net/~deadsnakes/+archive/ubuntu/ppa
[documentation]:
https://gregoryszorc.com/docs/python-build-standalone/main/technotes.html#how-it-works

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec