[query] Check for None/empty paths in vds combiner (#15000)
While running a combine job. I encountered a missing path which lead to
a very confusing set of errors where we were treating the missing path
as a hail assertion error rather than the data error that it was.

To fix this, change the getOrAssert calls in GVCF partition reader to
getOrFatal, causing a user error with missing data. Furthermore, add
checks to `new_combiner` to make sure that no paths are missing/empty
either.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP