[compiler] fix testBinarySearchOnDict (#14995)
## Change Description

See #14994.

Note that all the special casing of whether the dict contains the key is
unnecessary. We know the dict is sorted (we're using a binary search
after all), so if `dict(i).getKey < key` and `key <= dict(i).getKey`,
then if the dict contains the key, then `key == dict(i).getKey`.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP