[compiler] refactor extract aggregators (#14922)
## Change Description

Refactors the "extract aggregators" method to be (imo) much easier to
understand and use. This is the pass that takes an IR using `ApplyAggOp`
and the like, and extracts imperative IR (using `InitOp`, `SeqOp`, and
the like) for three separate phases of aggregation: initializing the
aggregator states, updating the states for each value in the collection
being aggregated, and computing the final value of the original IR from
the final aggregator state.

One of the sources of complexity in this pass is that some consumers
need those three phases to run in isolation, e.g. in `TableAggregate`,
or any distributed aggregation. This means any let bindings which are
used in multiple phases need to be duplicated. But some consumers, e.g.
`StreamAgg`, can simply define those bindings once and have them in
scope for the entire aggregation loop.

Previously, we were handling the later case with a very complicated
`rewriteFromInitBindingRoot` method, which I won't attempt to explain.
The new approach is much simpler: explicitly separate out any let
bindings which are accessed from the initialization phase (and which may
also be accessed in the final result phase). Then consumers who need the
phases to be isolated can prepend those bindings to both the
initialization and result IRs, and let `ForwardLets` remove any which
aren't used, while the other consumers can prepend those bindings once
to the entire aggregation process.

A few other simple refactorings better structure the code, and factor
out some frequently duplicated code.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP