[qob] Don't set hail off heap memory limits in QoB (#14978)
Also, set the java heap limit to 90% of machine memory.

The intention here was to give nicer error messages for when hail was
using too much memory. Unfortunately, this drastically reduces the jobs
that will succeed, as we partition the heap into a JVM portion and a
native portion. And so, the job may hit a memory limit while the machine
still has ample memory.

## Security Assessment
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating
- This change has low security impact

### Impact Description
This is a change to the environment of our java containers, where we run
Query on Batch jobs. As this is simply a change to how a container is
set up, this has low to no impact on the service as a whole.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec