[query] Use scalac-options library (#15027)
## Change Description

The main changes here are:

- Use the typelevel scalac-options library to more easily manage scalac
compiler flags, especially once we start maintaining a cross-version
build.
- Change how we specify build modes (previously debug vs release).
- Use language feature flags, instead of importing from
`scala.language`​ in every file where we use some gated feature like
higher-kinded types or implicit conversions. This is effectively the
only change to our compiler options.

### Build modes

There are now three build modes: Dev, CI, and Release:

|  | Verbose | Fatal Warnings | Optimize | Checked Allocator |
| --- | --- | --- | --- | --- |
| Dev | :white_check_mark: | :x: | :x: | :x: |
| CI | :x: | :white_check_mark: | :x: | :white_check_mark: |
| Release | :x: | :white_check_mark: | :white_check_mark: | :x: |

(“Verbose” is compiler flags like `-explaintypes`​ which enable more
verbose warnings and errors.)

To work around limitations of Mill/BSP/IntelliJ, namely that we can’t
set environment variables to control the IntelliJ build, I’ve split off
the meta-build in Mill, and have it check both an environment variable
and a file, as follows:

- If the `HAIL_BUILD_MODE`​ variable is set, use it,
- else, if the `hail/hail/mill-build/config/hail-build-mode`​​ file
exists, use it’s contents,
- else, default to Release.

Those of us using IntelliJ can create the file, which I’ve added to
`.gitignore`​, to use the Dev mode, and can still use the envvar to e.g.
build a release jar. Others can simply add the envvar to their shell
config.

This use of the Mill meta-build is really the right way to handle the
build mode anyways, since it’s an input that changes the structure of
the build itself (in particular, the source directories of the `memory`​
module). We were previously doing something a bit hacky to deal with
this.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP