[query] Daemonize Py4JQueryDriver http server
PR #14698 refactored the startup of the py4j http server so that calling
`pyHttpServer` created and started the sever. In the process, it deleted
the daemon thread that the server was started from originally. This
caused the `httpServer.start` method to spawn a non-daemon thread as it
was being called from the main thread py4j was using to do its work.
Simply calling `setExecutor(null)` is not sufficent for this purpose.

Manually stopping the server, like with `hl.stop()` would solve the
issue since the JVM would exit and spark would note that the job was
complete after the python process exited as well.

We once again create a thread that itself starts the http server, and
then start that thread in `pyHttpServer`.

This should solve the hanging in dataproc that bisection showed was
introduced in #14698.