[batch] Global security headers (#15037)
## Change Description

Fixes https://github.com/hail-is/hail-security/issues/37

- Replaces the annotation based implementation of API security headers
with standard global headers.
- Does not impact the CSP which is still implemented via a web security
header annotation
- Adds web security headers to the batch-driver UI pages


## Security Assessment

This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

Delete all except the correct answer:
- This change has a low security impact

### Impact Description

Replace this content with a description of the impact of the change:
- The effective change is that static content now also receives
`strict-transport-security` and `X-Content-Type-Options` headers.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec