[batch] Stay on batch page after cancelling a batch (#15044)
## Change Description

Fixes #14827

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

- This change has a low security impact

### Impact Description

The "Cancel" button on the batch page now redirects back to the details
page for that batch instead of the main batches list page.

When testing using the dev proxy, the cancel button would return a 405
error because the proxy was forwarding the cancel request as a GET
instead of POST. Changing `get` to `view` lets the request go through
correctly.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec

---------

Co-authored-by: Chris Llanwarne <cjllanwarne@users.noreply.github.com>
Co-authored-by: Christopher Vittal <cvittal@broadinstitute.org>
Co-authored-by: Patrick Schultz <pschultz@broadinstitute.org>