[query] initial support for cross compiling with scala 2.13 (#15029)
## Change Description

This PR adds support for building Hail with Scala 2.13. Though I don’t
have any immediate plans for how this could be used in production, it
has already proven very effective purely as a linter. And as the 2.12
compiler is effectively unmaintained at this point, it’s good to have as
few things keeping us stuck with 2.12 as possible.

The [Scala 2.13 change
log](https://github.com/scala/scala/releases/tag/v2.13.0) is a good
summary of all the changes. See also the [official
doc](https://docs.scala-lang.org/overviews/core/collections-migration-213.html)
on migrating to 2.13. Calling out those changes that most impacted the
initial migration:

- `Seq`​ is now an alias for `collection.immutable.Seq`​, and likewise
for `IndexedSeq`​. One consequence is we can no longer pass an `Array`​
where an `IndexedSeq`​ is expected. This is absolutely the right
default, but it does add a wrinkle to cross-compilation. I’ve left
`Seq`​ and `IndexedSeq`​ unelaborated, so they are immutable when
building with 2.13, and generic when building with 2.12. Of course, that
means that in practice, they are always immutable.
- As a consequence of the previous bullet, varargs are now exposed as
immutable `Seq`​s.
- Added a collection type `immutable.ArraySeq`​, which wraps an
`Array`​. This effectively does what we hackily tried to do with the
`FastSeq`​ stuff, in a much nicer way.
- Subclasses of `Growable`​ are expected to implement `+=`​ in 2.12, and
`addOne`​ in 2.13. I’ve added a shim trait `GrowableCompat`​ with
different implementations in 2.12 and 2.13 to bridge this inconsistency.
(Similarly for `Shrinkable`​.)
- Parallel collections are now in a separate module. In practice, this
means to use `someCollection.par`​ requires the import
`scala.collection.parallel.CollectionConverters._`​. Since this doesn’t
exist in 2.12, I added an empty object in the 2.12 only sources, and
suppressed the unused import warnings for it.
- Views have been redesigned and are more consistent and useful. The
most immediate impact is that Views are no longer subtypes of their
eager collection type.
- `Map.filterKeys`​ and `Map.mapValues`​ were weird, because they
returned views, unlike every other method on `Map`​. They’ve been
deprecated, and should be replaced with e.g.
`map.view.filterKeys.toMap`​. I’ve left all deprecations for later, but
where one of these methods was used where a (non-view) Map was expected,
I had to fix it now because of the previous point.
- Similarly, `zipped`​ has been deprecated, to be replaced with
`coll1.lazyZip(coll2)`​. There were a few places I had to make this
change now, because the wrong collection type was being inferred.

Critical to the cross-compilation story is the
`scala.collection.compat`​ library. It has different implementations in
2.12 and 2.13: the 2.12 version backports some of the new functionality,
and bridges between the old library and the new way of using
collections. The 2.13 version is mostly aliases to the standard library.
When building in 2.13, it often gets flagged as an unused import, so
again I’ve manually supressed those warnings.

I was also careful to individually silence each deprecation warning we
hit. This made the follow-up work of fixing deprecations far more
manageable, as we can just delete those silencers one at a time.

### Build system changes

The tldr recommended setup:

- If not using IntelliJ or Metals, the only change is needing to specify
a Scala version when invoking Mill directly, e.g. `./mill
hail[2.13].__.compile`. The make targets are unchanged.
- For IntelliJ, there are options described in detail below, but I
recommend making sure `SCALA_VERSION`​ is unset in your environment, and
creating the file `hail/hail/mill-build/config/scala-version​`, with
contents `2.13`, then resyncing BSP in IntelliJ. Everything should then
work fine as long as you only need to build with 2.13. To build with
2.12 in IntelliJ, just edit `scala-version` to `2.12`​, and resync BSP
in IntelliJ. You can also compile with either from the command line as
in the first bullet.

#### The details

This uses Mill's built in support for cross compilation. The `hail`
module is now parameterized by a Scala version, which may be one of
2.12, 2.12.13 (for our Azure build), or 2.13, with the minor versions
for 2.12 and 2.13 hardcoded in `build.mill`. At the command line, this
changes from `./mill hail.foo` to e.g. `./mill hail[2.12].foo`. The
build outputs similarly move from `out/hail/...` to `out/hail/2.12/...`.
This means build artifacts are kept for both 2.12 and 2.13, so switching
back and forth doesn't require rebuilding from scratch.

Unfortunately, for now, the IntelliJ/BSP/Mill system doesn't handle
cross versions well. The workaround I'm using is to use the Mill
meta-build to set the list of possible Scala versions, so that when
building in IntelliJ, Mill is configured to support only one Scala
version, and everything works fine.

The list of possible Scala versions is configured the same way as the
build mode (as of #15027). In particular:

- If the `SCALA_VERSION`​ variable is set, use it,
- else, if the `hail/hail/mill-build/config/scala-version​` file exists,
use its contents,
- else, all supported versions (2.12, 2.12.13, and 2.13) are enabled
(e.g. `./mill resolve hail[_]`​ shows all three). In either of the first
two cases, it will only show the specified version.

Additionally, note that unused imports warnings are disabled in 2.12.
The reason is that some imports are only used in one Scala version and
not the other. The 2.13 compiler has the ability to selectively disable
unused import warnings by a regex matching the import, making it easy to
disable those few cases. But 2.12 does not. As a consequence, using
Scalafix to remove unused imports must be done with 2.13, e.g. `./mill
hail[2.13].__.fix`.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP