[query] enable a few easy 2.13 warnings (#15031)
## Change Description

Enables the following 2.13 warnings:

- Dead code - when an expression is inferred to have type `Nothing`​,
anything using that value is dead code
- "export" deprecation - "export" is a keyword in Scala 3, so in 2.13 it
is deprecated to use unless wrapped in backticks
- match analysis - unused match cases
- implicit types - implicit declarations must have explicit types
- legacy bindings - In 2.13, inside a class defined inside an outer
scope, it is deprecated to refer to a name ambiguously bound in both the
outer scope and a parent class or trait. The fix is just to make it
explicit which is being referred to.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP