[query] `HailContext` is not required in batch (#15101)
Removes creation of a `HailContext` for QoB. `HailContext.apply`
previously registered a dense matrix operation. For compatibility, this
has been moved to a `lazy val` in the newly added `linalg` package
object (so that it's registered at most once). Each backend initialises
that `lazy val` in a similar way to the `StreamReadConstraints`
business.

This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP

Fixes #15072