[query] Don't filter entries with missing reference GT in combiner
The expression `hl.case().when(missing_expr, value).or_error(message)`,
evaluates to missing. This is not what we want in the combiner, where we
wish to error if there is a reference block with a non-ref GT, but
a missing GT is fine as it isn't non-ref.

Add a missingness check to `make_ref_entry_struct` in the combiner to
ensure that data in reference rows with missing GTs are preserved.