[query] vanquish `HailContext` (#15104)
Removed the `HailContext` class and moved its functionality to more
appropriate locations. The Java version check was moved to
`Logging.preamble()`, which now displays the Hail version and warns if
not using Java 11. The grep functionality was moved to
`Py4JQueryDriver`.

This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP