[auth] Scope down project-level ServiceAccountUser role for batch user (#15139)
## Change Description

Fixes https://github.com/hail-is/hail-security/issues/40

The batch user has the ServiceAccountUser role binding at a project
level, which grants it access to any service account in the project.
This change removes that binding & replaces it with bindings to each new
individual Hail user account. As part of that, we will first need to run
this script to backfill permissions for old existing users.

## Security Assessment

Delete all except the correct answer:
- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

Delete all except the correct answer:
- This change has a medium security impact

### Impact Description

Permissions for the batch user are scoped down. To do so, the auth user
now needs permission for `iam.serviceAccounts.setIamPolicy`. I tested
both these changes in a sandbox environment & did not see any impact
when creating users or running jobs.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec