[query] Lower BlockMatrixAgg with BlockMatrixStage2 (#15095)
This is a near direct port of the lowering to BlockMatrixStage. In the
both axes case (`bm.sum()`) we reduce each block to a single element,
before collecting and summing those elements. In a single axis case (ex.
`bm.sum(axis=0)`), we group by row/column block in a single job to
perform the aggregation, rather than a multi-stage approach, reducing
each block to a vector, then grouping and reducing each row/column block
of column/row vectors to a single vector.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP