[services] Add http verb and query params to admin logs (#15144)
## Change Description

Fixes #14988 

Adds the http verb to accessible metadata for easier filtering. Also
adds query parameters in analogy to existing formdata logging. All
visible only to admins for general monitoring of the system.

## Security Assessment

- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

- This change has a low security impact

### Impact Description

Adds more content to the system logging that goes to gcloud. Visible
only to admins for monitoring and threat detection.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec