[auth] Fix login and simplify STS header process (#15167)
## Change Description

Fixes #15165 which was caused by 302s being caught and replaced by our
error handler middleware without properly including all of the session
management.

That middleware was introduced to make sure error responses still had
strict-transport-security headers applied, but we can achieve the same
effect using the envoy proxy with much less complexity in each of the
services

## Security Assessment

- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

- This change has a medium security impact

### Impact Description

Replaces service-by-service STS header application in middleware with a
single point-of-exit header application in the envoy proxy. As a
side-effect this re-enables users to log in to the UI

### Appsec Review

- [ ] Required: The impact has been assessed and approved by appsec