[docs] First pass at some flaky test identification queries (#15156)
## Change Description

Some basic SQL to identify flaky test jobs, and success/failure
percentages.

Deliberately not trying to boil the ocean here. Very manual, lots of
false positives, but better than nothing. Hopefully we can add and
refine to this in the future, if they're useful.

## Security Assessment

This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP