[query] Change BlockMatrixType to always be 2 dimensional (#15151)
There was a lot of complexity in dealing with block matrix type as it
was possible for the shape of a 'matrix' to be zero or one dimensional,
in the case of a scalar or row vector. By making the type always have an
`nRows` and `nCols` parameter, I was able to remove
matrixShapeToTensorShape and tensorShapeToMatrixShape while simplifying
much of block matrix IR construction and type checking.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP