[query] Simplify BlockMatrix transpose IR generation (#15147)
BlockMatrix IR is a little wierd right now. We often try to treat
matricies with a single row/column as a vector. This leads to a lot of
additional complexity for little benefit. As as start to simplifying
this, always use `[1, 0]` as the index expression to
BlockMatrixBroadcast in transpose.

## Security Assessment
- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP