[query] engoodening of the py4j clients (#15158)
This change came about from getting the batch backend to run in a py4j
context.
I found the initialisation code in python quite busy and wanted to clean
it up.

The goal is to move as much functionality into the python `Py4JBackend`
so that it can accept any scala executor wrapped in the scala
`Py4JQueryBackend`
class. This change gets pretty close to that goal with the exception of
creating
and managing the lifetime of the jvm and spark context.

The main changes are as follows:

1. `SparkBackend`​ creates a `SparkSession` instead of a `SparkContext`.
Why?
- Because we need a `SparkSession` to convert to and from data frames in
python.
    - You can get the `SparkSession`'s underlying `SparkContext`,
- You cannot create a `SparkSession` with `SparkContext` directly, and
- you get warnings about reusing an existing `SparkContext` when doing
so.
2. Configures spark progress bar via `SparkConf`​ rather than hacking
the scala code.
3. Uses `SparkBackend.getOrCreate`​ from python to use idempotent
initialisation by default.
4. Moves logging configuration calls into python
    - I think this is cleaner
5. Warns when setting the local tmpdir when using the `SparkBackend`
instead of doing nothing
6. Better support for using an existing `SparkContext`
    - Don't stop a `SparkContext` that we didn't create
    - Use spark's job groups to cancel only those jobs in a stage

This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP