[Batch] Support over 64 CPU machine types (#14643)
Fixes #14644. Stops the worker from allocating more than 255 namespaces
even if the "CPUs per slot" calculation would otherwise allow it.

This change has a low impact on the Hail Batch instance as deployed by
Broad Institute in GCP. It adds a cap on a calculation that we were
already doing internally, in order to allow more CPUs and memory on
custom machine types