[infra] Add missing actAs permission to batch agent (#15190)
## Change Description

Fixes #15146 

We removed project-wide `serviceAccountUser` permissions from the batch
agent service account, but some APIs like `attach-disk` check if the
caller has this permission on itself. Adding the permission at the
account level fixes this.

## Security Assessment

- This change potentially impacts the Hail Batch instance as deployed by
Broad Institute in GCP

### Impact Rating

- This change has a low security impact

### Impact Description

This grants the batch agent permission to act as itself. The change is
already live in prod & this commit is just to update the Terraform
state. Imported the existing binding with `terraform import` & then
confirmed an empty diff with `terraform plan`.

### Appsec Review

- [x] Required: The impact has been assessed and approved by appsec