[query] use log4j-api-scala for string interpolation (#15181)
Readers are encouraged to read the log4j documentation on performance
considerations. This change is motivated by a number of the issues
discussed therein.

We do a lot of string interpolation in logs which isn't great as these
logging APIs use call by ref instead of call by name.

log4j-api-scala uses macros to transform string-interpolated log
statements into efficient java versions, making sure we're not building
strings needlessly.

Their doc suggests inheriting from `Logging`​ and using the protected
`logger`​ field. One benefit of doing things this way is that logs
contain the name of the enclosing class without analysing the call
stack. I find it much easier to find where a log statement was generated
within the source code as a result.

The trouble with this `Logging`​ trait is that the `logger`is not
reconfigurable for reasons I'm yet to fully understand. I think it has
something to do with specifying the class loader to log4j's
`LoggingFactory`​..? Anyway, I repurposed our `Logging`​ trait to
provide the protected `logger`​ member that is reconfigurable.

This change cannot impact the Hail Batch instance deployed by Broad
Institute in GCP