[compiler] Refactor BlockMatrix sparsity representations in type and lowering (#15163)
## Change Description

The bulk of the changes are in the files `BlockMatrixType.scala`,
`MatrixSparsity.scala`, and `LowerBlockMatrixIR.scala`. I suggest
starting with those, with the summaries below.

I acknowledge this is a big change containing a lot of non-trivial
logic. Please do ask for explanations and clarifications.

### BlockMatrixType and MatrixSparsity
`BlockMatrixType` has a `sparsity`, of type `BlockMatrixSparsity`, which
is either dense or a set of present blocks. In this PR I've moved the
sparsity representation to `MatrixSparsity`, which is a generic encoding
of the sparsity pattern of a sparse matrix. For a `BlockMatrixType`, the
sparsity will be a `nBlockRows` by `nBlockCols` `MatrixSparsity`.

Besides moving `BlockMatrixSparsity` to `MatrixSparsity`, and removing
references to blocks, the encoding of the sparsity is also streamlined.
Before, `BlockMatrixSparsity` held an array of coordinates of present
blocks, in an arbitrary order. Methods that need the present blocks in a
particular order (usually column major) would need to sort them, and we
also built a `Set` of present blocks to handle the `isPresent` query.

Now, `MatrixSparsity` (in the sparse case) enforces that the array of
present blocks is always in column major order. This simplifies some of
the logic, and lets us handle `isPresent` with binary search (similarly
for unions/intersections of sparsity, which now take advantage of the
ordering).

In addition, `MatrixSparsity` now knows its dimensions, which
`BlockMatrixSparsity` did not.

I've also gotten rid of all the complicated methods on the CSC
(compressed sparse column) encoding (e.g. `transposeCSCSparsity`).
Instead, I do all the transformations on the simple coordinate list
encoding, and convert to CSC as late as possible.

### LowerBlockMatrixIR
The changes in this file are mostly about a redesign of the
`BMSContexts` class. `BMSContexts` bundles together a runtime array of
context values (similar to the contexts for a `TableStage`), with a
representation of the sparsity pattern, which conceptually maps each
context value to the coordinates of its block.

Before, the sparsity pattern was also encoded by runtime values. In the
sparse case, this used runtime arrays `rowPos` and `rowIdx` using a CSC
encoding. Working with this encoding created a lot of non-trivial
runtime logic. But sparsity is completely known at compile time, so all
this runtime logic was unnecessary.

There are still cases where we need to work with a CSC encoding of the
sparsity at runtime. These cases are now handled by the class
`DynamicBMSContexts`, which is essentially the same as the old
`BMSContexts`, but now has a minimal interface of three methods.
Importantly, these methods are only about consuming a block matrix, not
transforming to a new one (like transpose).

The new `BMSContexts` now pairs a statically known `MatrixSparsity` with
a `DynamicBMSContexts`. The big change from before is that all methods
that produce a new `BMSContexts` now handle all the sparsity logic at
compile time, and simply embed the new sparsity in the IR using
literals. This typically requires us to also embed an array of ints
mapping from old to new positions in the contexts array, which we use at
runtime to reorder the contexts as needed.

## Security Assessment

- This change cannot impact the Hail Batch instance as deployed by Broad
Institute in GCP